+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+### Changed
+- Updated prompt to return symptoms
+
 ### Removed
 -   Removed the code where we does the string matching for disorders  
     user_input_in_lower_case = user_input.lower()